Make drain library more reusable
Move more functionality from the kubectl cmd to a package with
fewer dependencies.
